[lib][affine] Port timeline transformations to the zircon level.
Port the majority of the functionality of the media TimelineTransform
library down to the zircon level.  The library is not technically time
specific.  Instead it implements a specific representation of
one-dimentional affine transformations, so rename to libaffine in the
process.

This is all preparing to make use of the library functionality in the
kernel as part of the implementation of kernel clock objects.  Make
sure to use the same fundamental representation of the transformations
so that the core of the transformations maintain compatibility with
what is currently being used in the media code.

Also:
++ Write unit tests.  The old library didn't have any, and probably
   could use some.
++ Fix a few small edge case bugs which were found by inspection while
   thinking up test vectors.

Change-Id: I0841ac5f53b9f8bf77a7fd903c97c2f8af3c3df8