[e2e] Document src/tests vs. E2E tests.
Documents the relationship of end to end product tests and the
"tests" build targets in src/BUILD.gn and its sub directories.

The subtlety is that end to end product tests must never be
included in build configurations of products they are not
intended to test, lest they would be automatically executed
by a try builder for such a product, and then fail. Therefore
it is better to exclude end to end product tests from the
area-global tests target, and only add them to product
configurations in //products.

Change-Id: I80cd1e727603f10415181d61f602ac1832dab47c