[escher][SCN-1444] Wait for mesh uploads to complete.
This fixes an issue where a "render command buffer" was
not waiting for a "transfer command buffer" to complete
because not all of the semaphores were waited-upon.

This was due to inherent finickiness of Mesh inheriting
from WaitableResource; it was easy to become confused
about whether to take the semaphores from the Mesh or
from the Buffers that it encapsulates.  After this CL,
Mesh no longer inherits from WaitableResource, making
this confusion impossible.

Testing: in addition to the usual tests, added a new test
to paper_shape_cache_unittests.cc which would have detected
this issue.

ES-104 #comment
ES-206 #comment
SCN-1444 #done

Change-Id: I3bc4032aff593c1568c527d2ee3ac5729345b74b