[ISP] Removing the ping/pong indicator from DMA Mgr
This is a breakdown of a larger change for adding DMA
functionality. This change just removes the bool which
indicates whether the DMA mgr is writing to the ping or pong
register block, as the register offset will be handled elsewhere.

The DMA module is not enabled. This has no functional changes.

Test: CQ. No functional changes, module is disabled.
Change-Id: Idca582c949af499c5c8d5a22ac12acc8183fa244