[ledger] Write references for downloaded pieces
To make this commit possible, the downloading logic is refactored. The
previous logic was a two-pass traversal which fetched only the required
pieces and then filled in a VMO by re-reading pieces from disk. The new
logic fills the VMO directly, by calling a new |GetOrDownloadPiece|
method that fetches missing pieces, writes them to disk if necessary
(with the correct references), and returns them immediately.

Special care is required for the index tree node pieces, that cannot be
written directly in |GetOrDownloadPiece| because they need the full
object available to extract their references. A |WritePieceCallback|
parameter is introduced to postpone their write-back.

Change-Id: I8a3baf43fa6113197f705cfc97f707fdf5a92a38
Bug: LE-697 #done
Tested: fx run-test ledger_tests
        fx shell run-test-component fuchsia-pkg://fuchsia.com/ledger_tests#meta/ledger_e2e_sync.cmx