[debugger] Show registers for non-top frames.
Now that registers are available in non-topmost stack frames, we can
support "regs" for arbitrary frames.

This implementation always requests the current CPU state when using the
top stack frame (to ensure nothing is out-of-sync) but shows the
unwinder-extracted values for non-topmost frames. There is a warning
when printing stale values.

The output of registers was indented by 2 spaces so the headings look
better.

Minor refactoring to better support commands that require frames.

Make Frame::GetGeneralRegisters() const.

Change-Id: I5a597594979361ea8b2a4dfd2a55531bbd41fc81