[iquery][inspect] Initial Health finding/formatting.
Creates a new mode for iquery that searches for health nodes.
Changes recursion to be a frontend for a depth variable that will tell
iquery how further down the hierarchy to recurse.

Create a FormatHealth interface for formatters and implemented the basic
version of the TextFormatter. The JsonFormatter is pending.

TEST=Updated formatters unit tests.

Change-Id: I11bb8c07dbd178d64dca5531ebcc89caa8701d0e