[fidl] Types with internal paddings need encoding and decoding.
Previously, we have been making an optimization in the C and LLCPP
bindings, where if the type has no handles and no out-of-line objects,
encoding and decoding can be skipped. This assumption is no longer true,
as the encoder zeros padding bytes using the coding table.

This CL calculates if a type (recursively) has padding, and informs the
C and LLCPP backends via the extended TypeShape and JSON IR.

Using the new criteria, encoding and decoding may be skipped if a type:
1. Does not have handles
2. Does not have out-of-line content
3. Does not have paddings

Change-Id: I8e36f970c5701bbf39d74b424fd4663455f40914