[debugger] Fix stack, add tsets.
The "stack" command did not save the input size and always passed 0 for
the size, which makes it not work. This hooks up the size.

Adds a test for the "stack" command. Minor updates to the existing
console unit test.

Convert the saved registers on the stack frame from debug_ipc structs
to the client API wrapper for a register which is higher level and has
more helper functions. This simplified the code that uses the registers.

Update the mock stack frame to take a vector<Register> instead of a IPC
StackFrame struct. The only thing used from the IPC struct was the SP
and the IPC register representation. It's simpler for tests to supply
toe raw SP and a vector of client Register classes. Most of the modified
files are just updates to users of the MockFrame constructor.

Update the memory analyzer to use the registers saved on each frame.
This will show where registers that were pushed on the stack used to
point to. Using the saved registers allows one asynchronous state to be
removed from the analyzer since they don't have to be requested from the
target.

Change the way the stack was annotated from "SP" to the CPU's register
for the stack pointer, and "BP" to "frame base" which is more
architecture independent.

Change-Id: I5225367ea54d131e40ae8d06545aa113ec9cfe2e