[System Monitor] add round-trip tests to dockyard_host
This CL adds manual testing to the dockyard_host to query the total device
memory and to gather a list of kernel object names (for jobs, processes, and
threads).

Also adds: total_runtime for each thread, RECENT request type to get the most
recent value in a Sample stream.

Jira: DX-1530
Bug: 56
Test: out/default/host_x64/dockyard_host
Change-Id: Ic160063b9b132748f531421e2a03756e60d525ba