[scripts][vim] Escape g:fuchsia_dir when matching
While not very likely that this will cause any issues, should it fail,
it could be quite confusing.  When we are trying to do a string match,
we should not use one of the arguments as a regex.  In Vim script this
means we need to escape the string as the only substring match through
regex.

Change-Id: I947ea3b0f1fb3ba309e5dcfb98225a71518367f6