[debugger] Look up type definitions.
Adds a new function to resolve forward definitions. This uses the index
to find a type declaration with the same name.

Adds a new helper function on ExprValue to look up a forward definition.
This also uses the existing Type helper functions to remove C-V
qualifications.

Add a test for the new functionality. It is not otherwise hooked up yet.

Change-Id: Ifc83a3569c69ee5b8fd2d062d66c617b78762891