[debugger] Add ExprEvalContext to formatter.
FormatValue used to take a SymbolDataProvider. This patch changes it to
an ExprevalContext which is higher level (it contains a
SymbolDataProvider inside of it). The reason is that we will later use
this context to resolve variable forward definitions using the symbol
index.

There is no behavior change. Using the ExprEvalContext beyond just
getting the SymbolDataProvider will be done in a future patch.

Change-Id: I531e965a0b5b46a78b66ef54f07d847bfa8195ab