[fuchsia-runtime] Add missing HandleType values, improve HandleInfo type
Add the remaining processargs handle type values to HandleType in
fuchsia-runtime and improve the HandleId type, now renamed to
HandleInfo, as required for the new Rust process launcher.

This type was originally going to be defined in the process_builder
crate itself, but it doesn't make sense to have separate definitions for
this in two crates.

Change-Id: I6750823f178e30bc11c9fabb5a9f8129754442df