[identity] Create `identity_common` for TaskGroup
`account_common`, the previous home for TaskGroup, is intended for
account_manager and account_handler specifically, (and not e.g.
token_manager, token_store, etc). Here we create a common location
`identity_common`, that all of identity can use.

Additionally, we are renaming the Context type (and related types), to
TaskGroup, to be more descriptive.

AUTH-167 #comment

Change-Id: I19fb25d197d873c6d2d8da73c08b298b0d152027