[zircon][dev][goldfish] Add flags to goldfish benchmark
This makes it possible to specify the tests to run by this
benchmark using command line arguments. First argument is
the size of each data transfer, second argument is the
number of transfers to perform. Third/forth argument
specific size/transfers for the next test, etc.

The default tests have also been adjusted to not take too
much time.

Test: goldfish-benchmark
Change-Id: I6fb634232f3639379edb7dc5287394130479645b