[zircon][dev][fvm]: Prevent race condition.
When the fvm_init thread is running it will eventually call
DdkMakeVisible and some other DdkOperations. If the device is unbound
before this happens, then it will create a race condition between
DdkRemove(which sets zxdev_ to nullptr) and any other Ddk* call.

To prevent this we allow unbind to be called once, and will block until
the thread(fvm_init) is completed. Then we allow unbinding to proceed.

Unfortuntely due to the current structure of the code, it is impossible
to deterministically assert that is not racy, only through inspection.
Now I'm working a document to refactor and cleanup the driver to make it
testable, and safer. This will be tracked as improving FVM test
coverage.

ZX-4144 #done

TEST=fvm-test

Change-Id: I5babf875ef1c57043c14b3f3446ce64c9b0df094