[dev][aml-usb-phy-v2] Refactor USB PHY code as a driver
This change moves the Amlogic USB PHY code for astro and sherlock
from a library linked into the board driver to a standalone driver.
This work is to prepare for OTG support
(dynamically switching USB between host and peripheral role).

This CL makes no functional changes to USB PHY configuration for astro.
For sherlock, we update the magic numbers for PLL PHY tuning to the
latest values from the Linux kernel.

To facilitate having the XHCI driver depend on the PHY initialization,
XHCI now optionally binds as a composite device driver against both
the PHY driver and its platform device node.
For now, this is only to enforce that the PHY driver loads first,
but in the future, the PHY driver will provide a protocol for services
provided to the controller drivers.

TEST: device-enumeration-test updated, paving stable for astro and sherlock.
Change-Id: Ice6f98a689ce04075374bc248fd158177f313041