[e2e] Select fyi builders more directly.
We indicate by a tag that a test should be run not on
try builders but only on an fyi builder. This can be
expressed statically by always applying the tag only to
the desired environment, rather than to dynamically
apply the tag to all environments based on inference
which environment we are in.

Also apply the same pattern to the sl4f test configuration.

Change-Id: I63dbb990d7e094409223aef01489f1b646f18410