[fidl] Generate coding tables for enums and bits
Doing so provides the necessary size information for the linearizer to
work with enums and bits in envelopes. This allows LLCPP bindings to
pass the fidl_compatibility_test.

FIDL-645 #done

In preparation of future work to validate enums and bits using
coding tables, we are always generating these coding tables, regardless
if in an envelope or not.

Change-Id: I0fd19ab9facd65ee36a6c475e7cdf22b26757248