[zircon][oom] Outside of bringup, reboot on OOM
When not in a development configuration, instead of trying to finesse
which task(s) to kill, simply perform an orderly reboot.

This is unfortunately not possible to test automatically currently. The
manual test is to verify correct behavior when using:

    $ k oom lowmem

which simulates an out-of-memory situation.

ZX-3637 #comment [zircon][oom] Outside of bringup, reboot on OOM

Change-Id: I779418fcb9ac3edaa8a1e087eeeb7e5eac76e144