[kernel][x86] Move x64_platform_tests -> arch/kernel/x86
x64_platform_tests were a small set of unit tests of the read_msr/write_msr
interfaces; move them from common zircon test directory/build targets to the
x64 ones.

Building the tests on ARM was not useful; they were mostly if-def-ed out;
and a subsequent CL would like to add a dependency on the CPUID library,
which is x86-only.

Change-Id: If5158bb8b9ca713f5111227f2b2020bee9be0cd8