[kernel][x86] Intel: Enumerate MDS vulnerability
MDS ('Microarchitectural Data Sampling') is a recently-disclosed speculative
execution information leak affecting many x86 CPUs. Enumerate the leak; it is
present on most Intel CPUs except Atom 'Bonnell/Saltwell' and 'Goldmont/
Goldmont+' CPUs and on forward CPUs with MDS_NO set in a capability register.

Enumeration is in the same style as Meltdown and L1TF.

Unlike Meltdown and L1TF, MDS is a small family of vulnerabilities; enumerate
MDS when any variant from the family is present.

ZX-4111

Tested:
$ k cpu features
[00010.015] 01219:01240> Vendor: Intel
[00010.015] 01219:01240> Microarch: Haswell
[00010.015] 01219:01240> F/M/S: 6/3c/4
[00010.015] 01219:01240> patch_level: 0
[00010.016] 01219:01240> Brand: Intel Core Processor (Haswell)
[00010.016] 01219:01240> Features: fpu sse sse2 sse3 ssse3 sse4.1 sse4.2 mmx fxsr xsave aesni clflush smep smap erms
[00010.016] 01219:01240> syscall nx rdtscp hypervisor
[00010.016] 01219:01240> Properties: meltdown l1tf mds

Change-Id: I51cc3958b11c59ac0c1067a36562b0f91a424f9d