[media][tests] Audio analysis lib functions for phase
This test-only CL adds functions to the audio analysis library,
specifically to measure phase response. This will be done during
Fourier transforms that we already perform in order to measure our
frequency response and SINAD.

Phase response will be increasingly important as we add more
advanced resampling (and other audio functions) to the media
subsystem.

Test: 7 new cases added that test the analysis lib.

Change-Id: Ic5e1225a9a7eb40c40007a2f9c6ef4af63ed98e0