[zxcrypt][fvm] Reenable test, add logging
We disabled this test when it flaked in CQ once.  We haven't seen it
fail again, haven't been able to reproduce it locally, and have merged a
variety of patches that might interact with the behavior here.  Reenable
the test and add some logging to give us a touch more information if the
flake were to reoccur.

FLK-251 #comment

Change-Id: Ib2a4316b2c0a247ab42f5a8f0f66b82f3f128a97