[cleanup][sl4f] Correct runTraceProcessor return value
This fixes a missing return lint triggered by an upcoming Flutter/Dart
roll. This method previously always returned true due to a missing return
in an async block. It now returns true if exitCode == 0; false
otherwise.

When process.exitCode == 0, the then() async block didn't end with a
return statement. Since we're already awaiting the result of the async
block (previously either false or null, though result was discarded),
I've refactored this to eliminate the then(), and simply await the
ProcessResult, from which we can check the exitCode synchronously.

Change-Id: I29e3fc9431580ccb7c3537ff6c81b3ea94b7980b