[netstack3] Send an ICMP response for unrecognized IPv6 Extension Header Options
BUG: NET-2245 #done

This change also undoes a change in the options parsing utility that enabled
not including the size of the options header (option type and option
length fields) in the option length value. This change was originally
done to support IPv6 extension headers' options parsing.

IPv6 extension headers' options seems to be more different from what the
options module originally supported so instead of updating the options
module (and all the implementations that depend on the options module
that don't need the same support) an extension header options specific
parsing utility was written to better take advantage of the common
format of the IPv6 extension header options and suit its error handling
needs.

Change-Id: I3fd8821f0c4237c59f7644a80d0af0f36627bf78