[ahci] Close possible race opportunity.
In PortComplete(), the port lock is released to call block_complete()
before the port state is updated. This potentially opens a window for
danger. Since the watchdog thread is only looking at non-completed
ops, this should be safe. However, let's close it to be sure.

Deferred transaction completion until after the port has been iterated
and transactions removed from the port state.

Bug: ZX-3922
Tests: blktest and fs-test.
Change-Id: Ib3cbb86f9f82729acf1e015f43c96bb8f989485e