[debugger] Add ExprEvalContext to casting functions.
This provides the ExprEvalContext to the casting functions so they can
call ExprEvalContext::GetConcreteType(). This will resolve forward
declarations in addition to the const-volatile stripping of the previous
code.

In general the behavior will remain unchanged unless casting with a
foward definition. The forward definition resolution is already tested
separately.

Change-Id: I42a677233f6f4737096c348dc10fad4c77367aaa