[fx] add list-packages and cleanup push-package
list-packages provides a command that users can run to list what packages
their build produces.
push-package is adjusted to be simpler and provide information based on
list-packages.
list-universe-packages.py is removed as it is no longer needed.
Calls to the module indexer are removed as it no longer has any clients.

Change-Id: Iaf92e06b0adaf6e7d6b655a140ae79e6f5b783cc