[fidl][llcpp] Use VisitInternalPadding to handle xunion padding
Previously, the xunion padding is zeroed in the generated code, and
validated in the walker. This change brings handling logic for that
padding much closer to union and struct paddings.

As a bonus, it fixes compilation error with GCC in Zircon, since GCC
does not allow [[maybe_unused]] on member fields, whereas Clang requires
it in this case.

Change-Id: I1d672f97bc395d5adf8a939d6b95a4891ac665f4