[graphics][compute] Update one struct member name.
An instance of `spn_vk` creates and controls all access to
target-specific state including the pipelines and descriptor set
pools.

The spn_device struct had an spn_vk member with a name that was a
holdover from a prior implementation.

This improves the name and is part of a larger effort.

Also, additional documentation was added to the top of `spn_vk.h`.

Testability: fx build/src/graphics/lib/compute:tests
             fx shell run spinel_vk_context

Change-Id: Icbbb0f1a81efa2bbff2b38989008086a93136cc5