[identity] Integrate TaskGroup with Account system
Here, we assign and use TaskGroup objects in a way that allows us to
terminate and remove objects gracefully and without risk of data
races.

Account, Persona, and TokenManager each have their own task group, keeping
track of tasks that use them.
(In the future we will have a task group for each persona, so that
we can tear one down and delete it without affecting the others).

For now, we intercept cancellation fairly coursely, whenever we wait for
a new FIDL message. In the future, we can do this more granularly (for
instance waiting for a UI or network) with very little effort.

Additionally, refactored init_account to make it
play better with the parking_lot locks that AccountHandler uses.

Proper teardown flow, including account removal, comes in a subsequent
CL.

AUTH-167 #comment

Change-Id: I448e9fdd5838f14dcc171faa1c59788682df2eaa