[kernel][pc] Bitmap vector allocation
Replaces the pow2_range_allocator with a bitmap allocator for
allocating interrupts. The pow2_range_allocator has an internal Mutex,
and performs heap allocation, which triggers a LockDep violation as
taking a Mutex whilst holding a SpinLock is a bug. At the point we
are calling p2ra are holding both the InterruptManager lock_ as well
as the InterruptDispatcher spinlock_.

As x86 only has 256 vectors, the statically allocated bitmap is small,
and can be traversed in 4 64-bit comparisons. Performing vector
allocations is therefore expected to be as fast or faster than p2ra,
although this is not benchmarked.

Change-Id: If475e6922ac354e9ba2aeb65eb813fcbb65eb066