[escher] Add simple bitmap font.
The goal is to blit the frame number to the screen,
in order to correlate with traces/logs while debugging.

Much of this CL is to create Escher test infrastructure
for reading back pixels from the framebuffer.  Scenic's
infrastructure is unsuitable for this purpose because
Escher unit tests run both on host (i.e. Linux) and
device.

ES-208 #comment Added digits 0-9 and some punctuation characters.
ES-209 #done

Change-Id: I1e5e8c75e1b9a639611852e43ca21a0001cfa1ce