[kernel][object] Unregister interrupt out of lock
InterruptDispatcher::Destroy previously performed the mask and
unregister whilst holding its spinlock_. As explain in the added
comment, this can cause a deadlock.

Performing these operations outside of the lock is safe as the lock is
not guarding these operations.

The deadlock scenario mentioned was observed whilst replicating
ZX-4257. With this change the deadlock and hang ceased to be
reproducible.

Change-Id: I1d52a5b434736576eb8404d926e2f5d3a9e3acc2