[componentmgr] Introduce ModelObserver
This CL introduces ModelObserver, a mechanism to synchronously observe
changes to the model. This CL introduces TestHooks, an implementation
of ModelObserver that observes updates to the Model in tests and
produces a string representing the component topology.

The topology string is stable in the sense that regardless of the order
in which component instances were added to the Model, the same string
would be produced for a given topology. In order to accomplish this,
AbsoluteMoniker and ChildMoniker implemnt Ord and PartialOrd.

Futhermore, each component instance has a name: the last ChildMoniker's
name in its path. This CL introduces a name accessor.

This CL uses TestHooks in tests and introduces a couple of new moniker
tests. This CL is a prerequisite for Hub v2.

Bug: CF-541

Change-Id: I2b41663ca0fac933c2643a55591192f470e1ecb1