[pci] Add test framework for PCI Protocol Drivers
This introduces a PCI driver test which creates a Fake PCI
bus driver, spawns a proxy devhost for a given BDF, and then
binds a protcol test driver to it. From here, tests can be run
within the test client exercising the protocol and have the
results channeled back to the test runner via the RunTests
protocol. Initial tests have been added to support testing
ConfigRead and ConfigWrite to prove the framework out.

This verifies the proxy build / loading from the previous CL.

ZX-3146 #comment

Change-Id: Ie3eeebbcc39549c8fcb382f45e5104dc49ecde6c