[benchmarking] Reorder benchmarks to fix slowdown of ledger perf tests
The run_scenic_benchmark based benchmarks do not make an effort to close
the graphics application being benchmarked at the end of the test.  Move
them to after the ledger benchmarks in attempt to reduce ledger
benchmark output noise (which recently started after the garnet and
peridot benchmarks merged).  The long term solution is to migrate the
graphics benchmarks to the e2e framework, tracked at PT-181 and PT-182.

PT-175 #comment

Test: peridot-x64-perf-dawson_canyon CQ passes

Change-Id: Ifeb197d7215333f133eb751423f970e475578eeb