[paver][system_updater] Move hack to paver from system_updater
The partition map is incorrect for astro so a hack was necessary to
ensure the correct partition was paved. This hack causes incorrect
behavior in system_updater when paving zedboot on non-astro targets.
Moving the hack to paver allows for us to apply it to just astro.

Also re-enable paving zedboot on chromebook.

Tested: OTA'd chromebook, paved astro, and `runtests -t paver-test`
Change-Id: I10572fe7ba078893477f638eb085344d10c6838e