[componentmgr] ModelObserver => Hook
This CL just renames some things. ModelObserver is now Hook. ModelObservers is
now Hooks. TestHooks is now TestHook. This is based on an offline conversation.
The motivation being that observer suggests that it has no side effects whereas
hook allows for the possibility of side effects. Long term, we intend to use
these hooks to probe and modify behavior within the component manager so rename
them to avoid confusion.

Bug: CF-541

Change-Id: If8d3ea6453d7b6035cd25a2b16935d9d4798ce1a