[mt8167][usb] Add missing USB-PHY configuration to host driver.
In the current cut of the mediatek host driver, there is some missing
PHY and regulator configuration which was previously being done by the
bootloader.  And updated bootloader yields a non-functional driver.
This adds the missing configuration to the driver to bring it back into
a working state.

ZX-3945 #done
Test: manual tests with a UMS and HID device.
Test: iochk -c 200 /dev/class/block/xxx

Change-Id: Idd64b43b1001986f87c13b88ff707703764a632c