[ISP] Adding ability to dump ISP registers
The debugging interface occationally needs to check
register values.  This CL adds the ability to dump
all the register values for the ISP configuration,
and a way for the test driver to retrieve those values.

Test: this CL includes a test using the ArmDeviceTester.
See zircon/system/dev/camera/arm-isp/test/TESTING.md
for instruction on how to run on-device tests.

Change-Id: Ie35be16570f70dd6fb672ddfb6fc00238c2c8563