[testing][gn] Fix assemble_system()s forwarding of metadata
Previously, it was accidentally forwarding the "metadata" key out of
the invoker's metadata scope. We actually want to forward, e.g.
test_runtime_deps which is a key *in* the metadata scope.

This fixes the recovery.zbi being missing from the swarming runs, which
in turn causes the recovery_simulator_tests to actually run. This went
unnoticed because of a separate problem where the test didn't notice
that QEMU was exiting quickly when it couldn't find recovery.zbi.

Before:

$ cat out/default/host_x64/gen/src/recovery/simulator/recovery_simulator_tests.deps.json
[
  "host_x64/test_data/qemu/qemu.tar.gz",
  "host_x64/test_data/qemu/multiboot.bin",
  "host_x64/test_data/qemu/target_cpu.txt"
]

After:

$ cat out/default/host_x64/gen/src/recovery/simulator/recovery_simulator_tests.deps.json
[
  "host_x64/test_data/qemu/qemu.tar.gz",
  "host_x64/test_data/qemu/multiboot.bin",
  "host_x64/test_data/qemu/target_cpu.txt",
  "recovery.zbi"
]

TEST-771 #comment [testing][gn] Fix assemble_system()s forwarding of metadata

Change-Id: If9c0a609962d214975523ba183e4785d0127542a