[componentmgr] Hub v2 Skeleton
This CL implements a Hub v2 skeleton:

Hub v2 is implemented as a Hook into the Model. It builds a
pseudodirectory as
the Model changes. It caches updates to the Model so that accesses to
the
Hub are efficient and do not require acquiring locks.

$ = a component instance directory

The following are implemented:

$/url
$/exec/resolved_url
$/exec/runtime/
$/children/<name>

More of the Hub will be implemented in subsequent CLs.

Subsequent CLs will also implement running Hub when component_manager
runs at startup.

Subsequent CLs will support different views into the Hub rooted at
different component instances.

Subsequent CLs will also support different projections of the component
instances on the system.

Bug: CF-541

Change-Id: I90a2374d271baa2862a497a93185470cbe2c8ca2