[rust_backtrace_info] Introduce library
This library will not actully output a backtrace but will instead output the
contextual information needed to properlly parse a backtrace. This will, if
set as the panic hook, allow zero guess work symbolization but the symbolizer
won't be usable until after we also update the backtrace crate as well.

Change-Id: Ia6889f007f0d437d94d673a4b8040935a19ba4cb