[fidl][llcpp] Remove use of unsupported template code
Support for break and continue were scheduled to be added to
text/template and html/template as part of Go 1.10. This was reverted
due to some security issues in html/template, and it wasn't seen as a
good idea to have support in text/template without support in
html/template. History on these issues can be found in:

https://github.com/golang/go/issues/20531
https://github.com/golang/go/issues/23683

This was used in a few templates effectively to filter out methods
fitting various criteria. To work around this, several new template
functions are added that return slices containing only the interesting
items for that section.

As this code was never incorporated into an actual Go release, the host
toolchain we're using to build Fuchsia cannot be updated as the build
will break. The workaround in this change allows us to continue with
that update.

Test: Running `fx build -k 0 tools/fidlgen_llcpp_zircon:update` does not
result in any changes to llcpp-generated fidl header files.

Change-Id: I7c210cab9edbde7d32276c17423bf82fdfb49ee6