[debugger] Move SymbolVariableResolver into SymbolEvalContext.
The SymbolVariableResolver was a helper class that did a portion of the
variable resolving pipeline. It was used in a few places, but mostly by
the SymbolEvalContext which used it to evaluate variable names.

There are a lot of similar-sounding classes and the implementations were
relatively small. More pressingly, the SymbolVariableResolver needs to
be able to resolve concrete type names which needs an ExprEvalContext,
which creates a reference counting cycle.

Merging these two classes removes the reference counting cycle and
simplifies things by having fewer similar sounding classes.

Converts some tests from the old file to the new one. The other tests
are duplicative so were not migrated.

There should be no behavior change.

Change-Id: Ib6c2ba0e453554b71ad14199a43786c9442739b4