[zircon][fs-test-utils]: Use IsolatedDevMgr.
In unit tests we tend to create a large number of block devices
(ramdisks). In order to prevent possible leaks the option to use an
isolated_devmgr was added to the fixture. By default it is false,
since the main intention of this fixture was to be used in performance
tests.

TEST=fs-test-utils-test

Change-Id: I12dc4008bc2ce93a9100518760a2b6435aac5651