[debugger] Fix suspend assertions.
Fix an assertion waiting for a thread to suspend. Previously the
notification of an exception could race with the suspend and trigger an
assertion. The new implementation checks for this case, and also detects
it faster than waiting for the full timeout as described in the added
comments.

In debug_agent's main, clean up the message loop on error. Otherwise
this will trigger an assertion on connection errors.

Removes a declaration for an unused function in SocketConnection.

Significantly lengthens the suspend timeout. The enulator on my fast
desktop was occastionally going beyond the 100ms timeout which causes
confusing behavior.

Change-Id: Ibefaf97e603e385cb77e74d039b5723114691610