[fidl][analyzer] don't resume native exceptions
Move the responsibility for resuming an exception from the analyzer's
OnNativeException() into the caller (crashsvc).

The original design for resuming from the analyzer assumed that the
analyzer would run asynchronously, but we've evolved to the point where
crashsvc always blocks on the analyzer so it's simpler to just resume
from crashsvc once the analyzer returns.

This also makes the analyzer protocol less directly tied to the
exception API, which makes it easier to transition to the new
channel-based exceptions without having to change the analyzer.

ZX-4031

Test: runtests (QEMU and NUC)
Change-Id: I07076177651d262286772f9b0c75e121781ce15b