[process_builder] Add test for statically linked PIE loading
This adds a test helper that is statically linked and which implements
just enough ELF parsing and dynamic symbol lookup to be able to look up
syscalls from the vDSO. It looks up the zx_channel_read and
zx_channel_write syscalls and uses them to parse the bootstrap message,
find a PA_USER0 channel handle in it, and read a message and echo it
back. The test uses this echoed message to verify that it loaded the
process correctly.

CF-661 #progress

Change-Id: Idc4534ab5f7bcf8498f0acb206f9339f28ff6816