[scenic] Enforce View Bounds
This feature clips meshes based off of the bounding box of the view to which
they belong.

Right now this feature is optional. If you want this feature enabled, all you have to do is set the build flag scenic_enforce_view_bound_clipping=true.

Testing - There are 3 new tests.

First unit test checks to see if setting view properties on a view
holder updates its clip planes. It does this for various different
bounding box configurations.

Second unit test in view_clipping_unittest.cc creates a view
with a rounded rect whose bounding box only partially overlaps that
of the view holder's bounding box. The clipped shape that is generated by
the PaperShapeCache during normal scene-traversal/rendering is compared against
a clipped shape that is explicitly created.

The third test is a pixel test which has a view with a rect that extends
beyond the view's bounds and takes a screenshot of that scene, after
which it checks pixel values to make sure they are what you'd expect.

SCN-1180 #comment.
ES-210 #done

Change-Id: I9cea2ec109a0c7fc8e55882db70da8867d553dd8