[oom] Partial revert of 6601487da (reboot on oom)
This reverts reboot-on-OOM in favour of previous behaviour (which was to
kill jobs with the kill-on-oom bit set). This is a temporary reversion
to avoid file system corruption until journalling is rolled out for key
file systems.

ZX-2093 #comment [oom] Partial revert of 6601487da (reboot on oom)
ZX-4277 #comment [oom] Partial revert of 6601487da (reboot on oom)

Change-Id: I0b2cab69715ecc71995e395939fbc6465b6d7858