[voila] add an e2e test using SL4F
This patch adds an e2e test that verifies that Voila successfuly embeds
a replica - that is, that a session shell ends up being displayed in the
embedded view. This verifies the correctness of view embedding across
Voila -> Sessionmgr -> test shell, allowing us to catch breakages such
as LE-746.

It consists of the following elements:

 - Voila gets parametrized so that we can use only 1 replica and the
   specific minimal test shell for the test
 - the test shell is modified to display big green circle in the middle
   of the screen
 - the e2e test verifies tha the central pixel of the screen is the
   right shade of green

LE-750

Change-Id: I3bb89b291097788602baadba5a888d97d9798905